pyproject.toml
[tool.poetry]
name = "openapi-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chromadb = "^0.4.24"
openai = "^1.17.0"
wget = "^3.2"
numpy = "^1.26.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
